pyproject.toml
[tool.poetry]
name = "django-test-task"
version = "0.1.0"
description = ""
authors = ["David-Roklem <roklem@mail.ru>"]
readme = "README.md"
packages = [{include = "django_test_task"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.1"
psycopg2-binary = "^2.9.6"
python-dotenv = "^1.0.0"
social-auth-app-django = "^5.2.0"
django-debug-toolbar = "^4.1.0"
pytest-django = "^4.5.2"
mixer = "^7.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
